+++ R/stats-kmeans-tidiers.R
@@ -2,7 +2,7 @@
 #' @template title_desc_tidy
 #'
 #' @param x A `kmeans` object created by [stats::kmeans()].
-#' @param col.names Dimension names. Defaults to `x1, x2, ...`
+#' @param col.names Dimension names. Defaults to the names of the variables in x.  Set to NULL to get names `x1, x2, ...`.
 #' @template param_unused_dots
 #'
 #' @evalRd return_tidy("size", "withinss", "cluster")
@@ -13,7 +13,11 @@
 #' @export
 #' @seealso [tidy()], [stats::kmeans()]
 #' @family kmeans tidiers
-tidy.kmeans <- function(x, col.names = paste0("x", 1:ncol(x$centers)), ...) {
+tidy.kmeans <- function(x, col.names = colnames(x$centers), ...) {
+  
+  if(is.null(col.names)){
+    col.names <- paste0("x", 1:ncol(x$centers))
+  }
   ret <- as.data.frame(x$centers)
   colnames(ret) <- col.names
   ret$size <- x$size