+++ NEWS.md
@@ -25,6 +25,7 @@ in the augment method for the chi sq test, .residuals column was renamed to .res
 - Import `tidy()`, `glance()` and `augment()` generics from `modelgenerics`
 - Added new vignette detailing use of `modelgenerics` and `modeltests` packages
 - Added `data` argument to `augment()` generic
+- tidy.kmeans now defaults to using variable names in output columns
 
 ## Deprecations